Refactor: Renamed `database.py` to `auth_database.py` and added explicit environment variable loading.
