+++ backend/api/auth.py
@@ -6,7 +6,7 @@ from sqlalchemy.orm import Session
 from datetime import datetime, timedelta
 from backend.models import User
 from backend.utils.security import hash_password, verify_password, validate_password
-from backend.database import get_db  # Function to get SQLAlchemy session
+from backend.auth_database import get_db  # Function to get SQLAlchemy session
 from pydantic import BaseModel
 from typing import Optional
 import os
