+++ backend/api/clients.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 from backend.models import ClientAccount as Client, User
-from backend.database import get_db
+from backend.auth_database import get_db
 from backend.api.auth import get_current_user
 from pydantic import BaseModel
