+++ docs/project_structure.md
@@ -14,6 +14,7 @@ JedgeBot/
 │   │   │   ├── clients.py
 │   │   │   ├── main.py
 │   │   │   ├── request_utils.py
+│   │   ├── auth_database.py
 │   │   ├── broker
 │   │   │   ├── __init__.py
 │   │   │   ├── broker_api.py
@@ -39,7 +40,6 @@ JedgeBot/
 │   │   │   ├── data_fetcher.py
 │   │   │   ├── data_processor.py
 │   │   │   ├── setup_auth_db.py
-│   │   ├── database.py
 │   │   ├── execution
 │   │   │   ├── __init__.py
 │   │   │   ├── orders.py
